workaround https://github.com/mapnik/mapnik/issues/2112

+++ osx/scripts/build_mapnik.sh
@@ -3,12 +3,15 @@ set -e -u -x
 
 cd ${MAPNIK_SOURCE}
 
-echo 'Building mapnik'
+echoerr 'Building mapnik'
 
-rm -rf ${MAPNIK_BIN_SOURCE}
-rm -f src/libmapnik{*.so,*.dylib,*.a}
-rm -f tests/cpp_tests/*-bin
-make clean
+if [ -d ${MAPNIK_BIN_SOURCE} ]; then
+  rm -rf ${MAPNIK_BIN_SOURCE}
+  rm -f src/libmapnik{*.so,*.dylib,*.a}
+  rm -f tests/cpp_tests/*-bin
+  # TODO: https://github.com/mapnik/mapnik/issues/2112
+  make clean
+fi
 
 if [ "${TRAVIS_COMMIT:-false}" != false ]; then
     if [ $UNAME = 'Darwin' ]; then
